+++ static/src/javascripts/bootstraps/enhanced/main.js
@@ -178,23 +178,6 @@ const bootEnhanced = (): void => {
         );
     }
 
-    if (!config.get('page.isHosted')) {
-        fastdom.read(() => qwery('.youtube-media-atom')).then(els => {
-            if (config.get('isMedia') || els.length) {
-                require.ensure(
-                    [],
-                    require => {
-                        bootstrapContext(
-                            'media',
-                            require('bootstraps/enhanced/media/main').init
-                        );
-                    },
-                    'media'
-                );
-            }
-        });
-    }
-
     if (config.get('switches.enhancedMediaPlayer')) {
         fastdom
             .read(() =>
